feat: port index.js to TypeScript
