+++ eslint-formatter-table.d.ts
@@ -1,1 +1,4 @@
-declare module 'eslint-formatter-table';
+declare module 'eslint-formatter-table' {
+  function formatter(results: unknown[]): string;
+  export = formatter;
+}
