+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "eslint-formatter-ratchet",
   "version": "3.0.0",
-  "main": "index.js",
+  "main": "dist/index.js",
   "repository": "https://github.com/Jmsa/eslint-formatter-ratchet.git",
   "description": "Ratcheting applied to ESLint results so new issues don't creep in.",
   "author": "James Abercrombie <jmsabercrombie88@gmail.com>",
@@ -15,12 +15,15 @@
   "files": [
     "README.md",
     "LICENSE.txt",
-    "index.js"
+    "MIGRATING.md",
+    "dist/index.js"
   ],
   "scripts": {
-    "test": "FORCE_COLOR=false nyc mocha ./index.test.js",
-    "test:ci": "FORCE_COLOR=false nyc --reporter=lcov mocha ./index.test.js",
-    "release": "standard-version"
+    "build": "tsc",
+    "build:watch": "tsc --watch",
+    "test": "yarn build && FORCE_COLOR=false nyc mocha ./index.test.js",
+    "test:ci": "yarn build && FORCE_COLOR=false nyc --reporter=lcov mocha ./index.test.js",
+    "release": "yarn build && standard-version"
   },
   "husky": {
     "hooks": {
@@ -55,7 +58,6 @@
   "dependencies": {
     "chalk": "4.1.2",
     "deep-object-diff": "1.1.9",
-    "eslint-formatter-table": "7.32.1",
-    "node-emoji": "2.2.0"
+    "eslint-formatter-table": "7.32.1"
   }
 }
