+++ tsconfig.json
@@ -10,6 +10,6 @@
     "forceConsistentCasingInFileNames": true,
     "declaration": false
   },
-  "include": ["index.ts"],
+  "include": ["index.ts", "eslint-formatter-table.d.ts"],
   "exclude": ["node_modules", "dist", "vitest.config.ts"]
 }
